Add Depth of field effect
and using double to improve precision

+++ common/hitable.hpp
@@ -8,7 +8,7 @@
 #endif
 class material;
 struct hit_record{
-    float t;
+    double t;
     vec3 p;
     vec3 normal;
     material* mat_ptr;
@@ -18,7 +18,7 @@ class hitable{
 
 public:
 //pure virtual method ,must be children class overwritten
-virtual bool hit(const ray& r ,float t_min,float t_max,hit_record& rec)const =0;
+virtual bool hit(const ray& r ,double t_min,double t_max,hit_record& rec)const =0;
 
 
 };
